Consolidate MoE routers
