+++ docs/Implementation.md
@@ -13,6 +13,8 @@ This repository includes starter modules for the first two algorithms listed in 
     `active_experts` to track the current count.
   - `RLMoERouter` learns routing probabilities with a REINFORCE update and
     mirrors the `ElasticMoERouter` API for plug-and-play experiments.
+    Both classes now live alongside `HashRouter` and `SwitchRouter` in
+    `src/moe_router.py`.
 - `src/moe_layer.py` implements a small MoE feed-forward block using these routers. It accepts an optional
   `balance_weight` which multiplies the `balance_loss()` penalty derived from the router's assignments and
   returns it alongside the layer output.
@@ -756,7 +758,8 @@ python scripts/attention_analysis.py --model model.pt --input sample.txt --out-d
 - **Implemented** an `ElasticMoERouter` that scales the number of active experts
   according to real-time GPU utilization.
 - **Implemented** an `RLMoERouter` that trains routing weights via a simple
-  reinforcement learning loop for improved load balance.
+  reinforcement learning loop for improved load balance. Both routers now reside
+  in `src/moe_router.py` for ease of reuse.
 - Extend `HierarchicalMemory` with an `SSDCache` that prefetches high-frequency
   vectors for faster retrieval. *Implemented with a disk-backed cache and
   persistence helpers in `src/hierarchical_memory.py`.*
