+++ docs/Plan.md
@@ -320,7 +320,7 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
 16. **Quantum architecture search**: Extend `QAEHyperparamSearch` to explore
     novel transformer components and report promising variants.
     *Implemented in `src/quantum_hpo.py` with unit tests.*
-17. **Elastic mixture-of-experts routing**: *Implemented in `src/elastic_moe_router.py`.*
+17. **Elastic mixture-of-experts routing**: *Implemented in `src/moe_router.py`.*
     The router varies active expert counts based on GPU load and compares load
     balance with the static `SwitchRouter`.
 18. **SSD-backed retrieval cache**: Extend `HierarchicalMemory` with an
