+++ src/moe_router.py
@@ -1,6 +1,7 @@
 import torch
 from torch import nn
 from abc import ABC, abstractmethod
+from typing import Callable
 
 
 class BaseRouter(nn.Module, ABC):
@@ -81,6 +82,103 @@ class SwitchRouter(BaseRouter):
         return torch.bincount(assignments.view(-1), minlength=self.num_experts)
 
 
+class ElasticMoERouter(SwitchRouter):
+    """Switch router that reduces active experts when GPU utilization is high."""
+
+    def __init__(
+        self,
+        dim: int,
+        num_experts: int,
+        k: int = 2,
+        temperature: float = 1.0,
+        min_util: float = 0.5,
+        max_util: float = 0.9,
+        utilization_fn: Callable[[], float] | None = None,
+    ) -> None:
+        super().__init__(dim=dim, num_experts=num_experts, k=k, temperature=temperature)
+        self.min_util = min_util
+        self.max_util = max_util
+        self.utilization_fn = utilization_fn
+        self.active_experts = num_experts
+
+    def gpu_utilization(self) -> float:
+        """Return current GPU memory utilization ratio."""
+        if self.utilization_fn is not None:
+            return float(self.utilization_fn())
+        if torch.cuda.is_available():
+            alloc = torch.cuda.memory_allocated()
+            total = torch.cuda.get_device_properties(0).total_memory
+            return float(alloc) / float(total)
+        return 0.0
+
+    def _adjust_active_experts(self) -> None:
+        util = self.gpu_utilization()
+        if util >= self.max_util:
+            scale = 0.25
+        elif util >= self.min_util:
+            scale = 0.5
+        else:
+            scale = 1.0
+        self.active_experts = max(1, int(self.num_experts * scale))
+
+    def forward(self, x: torch.Tensor):
+        self._adjust_active_experts()
+        assignments, weights = super().forward(x)
+        if self.active_experts < self.num_experts:
+            assignments = assignments % self.active_experts
+        return assignments, weights
+
+    def load_balance_std(self, assignments: torch.Tensor) -> float:
+        counts = torch.bincount(assignments.view(-1), minlength=self.active_experts).float()
+        return (counts.std() / counts.mean()).item()
+
+
+class RLMoERouter(SwitchRouter):
+    """Switch router that learns routing probabilities with REINFORCE."""
+
+    def __init__(
+        self,
+        dim: int,
+        num_experts: int,
+        k: int = 2,
+        temperature: float = 1.0,
+        min_util: float = 0.5,
+        max_util: float = 0.9,
+        utilization_fn: Callable[[], float] | None = None,
+        lr: float = 0.1,
+    ) -> None:
+        super().__init__(dim=dim, num_experts=num_experts, k=k, temperature=temperature)
+        self.min_util = min_util
+        self.max_util = max_util
+        self.utilization_fn = utilization_fn
+        self.lr = lr
+        self.baseline = 0.0
+        self.prefs = torch.zeros(num_experts, dtype=torch.float32)
+
+    def forward(self, x: torch.Tensor):
+        probs = torch.softmax(self.prefs / self.temperature, dim=0)
+        batch, seq, _ = x.shape
+        flat_probs = probs.repeat(batch * seq, 1)
+        flat_assign = torch.multinomial(flat_probs, self.k, replacement=True)
+        assignments = flat_assign.view(batch, seq, self.k)
+        weights = probs[assignments]
+
+        reward = -self.load_balance_std(assignments)
+        adv = reward - self.baseline
+        self.baseline += self.lr * adv
+        counts = torch.bincount(assignments.view(-1), minlength=self.num_experts).float()
+        freq = counts / counts.sum()
+        for i in range(self.num_experts):
+            grad = adv * (freq[i] - probs[i])
+            self.prefs[i] += self.lr * grad
+
+        return assignments, weights
+
+    def load_balance_std(self, assignments: torch.Tensor) -> float:
+        counts = torch.bincount(assignments.view(-1), minlength=self.num_experts).float()
+        return (counts.std() / counts.mean()).item()
+
+
 def balance_loss(assignments: torch.Tensor, num_experts: int) -> torch.Tensor:
     """Return relative std of expert usage as a loss term."""
     counts = torch.bincount(assignments.view(-1), minlength=num_experts).float()
@@ -104,3 +202,15 @@ def token_drop_rate(assignments: torch.Tensor, num_experts: int, capacity: int) 
     counts = torch.bincount(assignments.view(-1), minlength=num_experts).float()
     dropped = torch.clamp(counts - capacity, min=0).sum()
     return (dropped / assignments.numel()).item()
+
+
+__all__ = [
+    "BaseRouter",
+    "HashRouter",
+    "SwitchRouter",
+    "ElasticMoERouter",
+    "RLMoERouter",
+    "balance_loss",
+    "balance_loss_probs",
+    "token_drop_rate",
+]
